Fixed several issues with session context - now declared in the Resource instead of Service. Added ordering / sorting to medical record. Fixed missing @Stateless for several classes. Removed commented code. Fixed tests.

+++ tjeneste/src/main/java/no/arkivverket/helsearkiv/nhareg/medicalrecord/MedicalRecordConverter.java
@@ -238,7 +238,7 @@ public class MedicalRecordConverter {
             recordTransferDTO.setLnr(journalId.getLøpenummer());
         }
 
-        recordTransferDTO.setFanearkid(Integer.parseInt(medicalRecord.getFanearkid()));
+        recordTransferDTO.setFanearkid(Long.parseLong(medicalRecord.getFanearkid()));
 
         final List<Lagringsenhet> storageUnitList = medicalRecord.getLagringsenhet();
         if (storageUnitList != null && storageUnitList.size() > 0) {
